Extract out `Data.PercentageSpec` from `Data.QuantitySpec`.
